Rename src file to change detected language
In order to fix the repository displaying that the code is ActionScript because of the .as extension, the source file's extension was changed to asm as it is assembly. Not x86 or MIPS or any commercially known assembly per se, but assembly for the P3 processor, supplied to us by the faculty during labs along with an emulator to develop this game without the need to use 'the real deal'.